Update AssetParams total to be BigInteger

+++ dotnet-algorand-sdk/V2/Algod/Model/Model.cs
@@ -14,6 +14,7 @@
 
 namespace Algorand.V2.Algod.Model
 {
+    using System.Numerics;
     using System = global::System;
 
     /// <summary>Account information at a given round.
@@ -232,7 +233,7 @@ namespace Algorand.V2.Algod.Model
 
         /// <summary>\[t\] The total number of units of this asset.</summary>
         [Newtonsoft.Json.JsonProperty("total", Required = Newtonsoft.Json.Required.Always)]
-        public ulong? Total { get; set; }
+        public BigInteger? Total { get; set; }
 
         /// <summary>\[un\] Name of a unit of this asset, as supplied by the creator. Included only when the name of a unit of this asset is composed of printable utf-8 characters.</summary>
         [Newtonsoft.Json.JsonProperty("unit-name", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
